Proper checks.

+++ install.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 
 TARGET_DIR="/usr/share/huya/xkb"
+XKB_DIR="/usr/share/X11/xkb/symbols"
 if [ ! -d $TARGET_DIR ]; then
   echo "Installing HUYA/XKB for the first time"
 else
@@ -10,18 +11,17 @@ else
 fi
 
 echo "Cloning the repository..."
+[ -d $XKB_DIR ] || exit 1
+
 git clone https://github.com/mudasobwa/huya-xkb.git $TARGET_DIR
-cd $TARGET_DIR
-[ "$1" = "ask" ] && export ASK="true"
-cd /usr/share/X11/xkb/symbols || exit 1
-if [ ! -f ru_es ] ; then
-  ln -s $TARGET_DIR/ru_es.layout ru_es
-fi
+[ -f $XKB_DIR/ru_es ] || ln -s $TARGET_DIR/ru_es.layout $XKB_DIR/ru_es
+
+pushd $TARGET_DIR
 for i in ru es ; do
-  cp $i $TARGET_DIR/$i.bup
-  patch --dry-run $i < $TARGET_DIR/$i.patch
-  patch -b $i < $TARGET_DIR/$i.patch
+  cp -f $XKB_DIR/$i $TARGET_DIR/$i.bup
+  patch -b $XKB_DIR/$i < $TARGET_DIR/$i.patch
 done
+popd
 
 echo "Removing xkb cache..."
 rm /var/lib/xkb/*xkm 2>/dev/null
